SERVER-92225 fix bazel order conversion list (#24498)
When ThinTargets was merged it removed the bazel targets from the scons
graph. The libdeps tool was updated to account for linking and
summarizing the dependencies, but did not update the libdeps graph
generation which is what the bazel order conversion list relies on. This
change updates the graph generation and list to print out the correct
bazel conversion list as well as add some checks to CI.