SERVER-91942 Update HashLookupUnwindStage to properly preserve unowned values (#24531)
A bug was discovered in BlockToRow::doSaveState() where unowned values
weren't always being properly preserved when saveState() as called
(fixed by SERVER-91512).

The SERVER-91942 ticket was opened to audit the rest of the SBE stages
and see if there were any similar bugs. Looking through the code for
each stage, I found that HashLookupUnwind::saveState() had a similar
problem.

This PR fixes HashLookupUnwind::saveState() so that it properly
preserves the slots from the outer child when necessary.