SERVER-92150 Add check in interleaved decompression that all decoders are exhausted (#24499)
Fixes a fuzzer failure found in BF-33927

The iterative decoder has a check to ensure that all decoders exhaust
together at the end of an interleaved mode, which the block decoder was
missing. This would not have had any effect on decoding of valid data
and only impacted whether or not invalid data was asserted out.

Tested locally and relevant fuzzer strings now pass.