SERVER-90625 Have QueryStats sumOfSquares not overflow for 64-bit integral types (#22950)
# Issue

SERVER-90625: the sum of squares queryStats metric
(`metrics_totalexecmicros_sumofsquares_delta`) is overflowing because it
is being calculated in a uint64. Since this metric is likely to overflow
because is measured in microseconds and it grows exponentially, we
should use a larger data type.

# Description

The Decimal128 supports 34 digits. We anticipate this conversion would
lose some precision, though we doubt it matters too much.

# Testing

See PR's last comment for patch links.