SERVER-92113 Resharding oplog fetcher should not write reshardProgressMark no-op to buffer if timestamp is minFetchTimestamp (#24436)
Doing so can result in a bug where the no-op entry never gets applied by
the resharding oplog applier because it only applies entries in the
buffer with a timestamp > minFetchTimestamp. This can then cause the
estimated oplog application time to be high and prevent resharding from
entering the critical section. See:
[BF-33742](https://jira.mongodb.org/browse/BF-33742) for more
information.