SERVER-80272 Consolidate logic to overflow oplog in tests (#24377)
SERVER-80272 is a fix for an old cold BF I didn't get around to until
now. In short, the way that too_stale_secondary.js overflowed the oplog
was not deterministic and in rare cases could cause the test to fail if
it took a very long time for the truncation thread to run. I did some
minor refactoring here to let the test reuse the much more deterministic
oplog rollover logic from oplog_rollover_test.js.