SERVER-92146 Rectify grouping of $topN/$bottomN by changing use of $ifNull for missing fields (#24595)
This is an update of a [previous
PR](https://github.com/10gen/mongo/pull/24514) which has been
[reverted](https://github.com/10gen/mongo/pull/24577).

Only the last commit "Adjust sortBy in integration test..." is new. Some
integration tests relied on the group _id "$s" as a sort key for $topN
accumulators, which resulted in an inconsistent ordering in a sharding
context (e.g. the `aggregation_sharded_collections_passthrough` suite).
Now, a different sort key is used, which should always produce
consistent ordering in the results (there is no ambiguity anymore).