SERVER-92035 Bound values for QE range parameters (#24566)
Adds validation that in QE range that the combination of sparsity,
trimFactor, precision (for floating point), min and max is reasonable.
It is possible to create situations where it would be impossible to do
some queries because the number of edges would not fit in a BSON
document. This change adds that validation and unit tests.