SERVER-92389 Make a new shared_ptr to avoid use after free (#24721)
[BF-34120](https://jira.mongodb.org/browse/BF-34120) was caused because
when the JWK background refresh job was running, a client called the
refreshKeys command at the same time.

This caused a `use after free` because the refreshKeys cmd deleted the 
underlying JWKManager object _keyManager was holding just after
JWKSetRefreshJob grabbed a reference of its keys to compare them later.

This solution creates a new shared_ptr so the ptr count does not go to 0
when the refreshKeys cmd is called and the underlying JWKManager obj
does not get destroyed.

I created a list of scenarios showing two columns Thread 1 and Thread 2,
showing how this is happening and what will happen with this solution.

https://docs.google.com/spreadsheets/d/1QOoIcxMI_vzsS-JmR-M150_a9Oy2oB_sIiMbti7Gy3w/edit?usp=sharing.

---------

Co-authored-by: Erwin Pe <erwin.pe@mongodb.com>