SERVER-91834:Investigate _shardsvrCoordinateMultiUpdate system user authenticated run (#24141)
https://github.com/10gen/mongo/blob/master/jstests/auth/internal_command_auth_validation.js#L561
has a test for _shardsvrCoordinateMultiUpdate. This PR enhances the test
for that command.
1) For a user without ['__system'] role, _shardsvrCoordinateMultiUpdate
will return an Unauthorized user error.
2) For a user with ['__system'] role, _shardsvrCoordinateMultiUpdate
should not return an Unauthorized user error.

In an effort to run this test the following changes are added
1) sys and norole users are now constant (this is needed for different
commands to have admin as user)
2) commands can have optional precommands to run. This precommands can
be run on a different server or db.
3) commands can have optional postcommands to clean up blocking work
from a command.