SERVER-90679 adding instructions on how to download mongot binaries from evergreen artifacts (#24864)
Adding a section to the mongot_testing_instructions that explains how
you can download the mongot binary that was used by a specific evergreen
build variant for a given commit or evergreen patch.

OLD PR: https://github.com/10gen/mongo/pull/24670
- Apologies, I had some weirdness happen with branches that I didn't
pick up on right away, but this is now appropriately titled and includes
all the latest changes, but I had to close the old PR. The only
un-resolved discourse on that PR was w.r.t. downloading from evergreen
versions, and I'm currently waiting for a response on that.