SERVER-91201: Add score fusion test to with_mongot e2e js tests (#24541)
[SERVER-91201](https://jira.mongodb.org/browse/SERVER-91201)

**Description**
Adds the `score_fusion_test.js` file that runs the analogous hybrid
search to the one in `ranked_fusion_verbose_test.js`, but mixes and
sorts the results with the score fusion algorithm. The normalization of
the vector and search score (before mixing) are done with the two
methods laid out in the technical design doc: Sigmoid and MinMaxScalar.

Also migrates the associated rank fusion test file to use the
`assertDocArrExpectedFuzzy()` fuzzy asserter.

---------

Co-authored-by: Kyra Ramesh Krishna <kyra.rameshkrishna@mongodb.com>
Co-authored-by: Will Buerger <59492746+wbuerger46@users.noreply.github.com>