SERVER-92645 Allow untimestamped writes for restore (#24995)
Some of the restore procedure requires booting in standalone mode which
can't accurately be checked at the stage in boot where untimestamped
writes are enabled. This ticket adds a specific override of this when
`--restore` and `--allowUnsafeUntimestampedWrites` are set on the
mongod.

Additionally, the server's tests shouldn't journal every collection
during the intermediate standalone restore boot. Hence the added usage
of `wiredTigerSkipTableLoggingChecksOnStartup`.

---------

Co-authored-by: Gregory Wlodarek <gregory.wlodarek@mongodb.com>