SERVER-92442 Make FeatureFlagUtil JS library retry making it's initial connection to the cluster on network errors (#24888)
Change the way that FeatureFlagUtil library obtains a connection to a
mongod in a sharded cluster by allowing it to use the connection created
by ReplSetTest. ReplSetTest will ensure that establishing the connection
is resilient to network errors/will retry in that case, and subsequent
use of the connection should be resilient to network errors when
applicable via runCommand overrides.

---------

Co-authored-by: huayu-ouyang <huayu.ouyang@mongodb.com>