SERVER-90893 Make list databases consistent to a given storage snapshot (#23788)
This change makes list databases use a storage snapshot, and read lock
free instead of taking a global lock. The catalog helper that lists the
database will first look at the committed list, then at the pending
list. If the storage snapshot considers the pending list as committed,
the database will be considered into the list, if not already there.