SERVER-92517 Fix TaskExecutorCursor creation for search explain (#24908)
Since there can be multiple mongots that a mongod connect to, the
`TaskExecutorCursor` (TEC) needs to do setup necessary to pin the cursor
to the mongot it initially queried to make sure that getMore's work
properly. This means that we need to call `makeTaskExecutorCursor()`to
handle the creation of the TEC. The newer version of mongot will work as
expected for this situation.

For older versions of mongot

- The TEC will query mongot and error on the response as it has no
cursor.
- We catch this error and create a dummy TEC ourselves.
- We will query mongot again for the explain object.

This code is temporary until SERVER-91594 can be done where we can
remove all the code related to this.

Alternatives considered
- Modify the TEC class to handle the error
This would require search explain specific changes to the TEC class, and
would not be ideal for easily removable code or for modularization of
code. The TEC class would have additional logic explicitly for this very
specific and temporary case.

- Do the setup required for pinning of the TEC within search code and
then pass the response back to the TEC
The code related to the setup of pinning the connection is complicated
and not easily splittable from the TEC. This will be hard to do and to
do correctly, especially purely within the search code.

Although catching a specific error and querying mongot twice is unideal,
it seems like the best option between the three in terms of isolatable
code that can be removed easily.