SERVER-92644 Add a background hook to resmoke that periodically adjusts runtime parameters of all nodes in a cluster (#24993)
Write a new background hook in resmoke that periodically calls
setParameter on every node in the cluster. Wire up a new command line
argument to resmoke to add this hook to the hook-list for a suite by
default when this parameter is specified.

This patch only adds the scaffolding and hook itself, deferring the
logic of selecting what parameters to update, what tasks to run in, what
constants to use for sleeping between iterations, etc, to future work.
Support for changing parameters on mongos and handling errors that may
arise on kill-primary configurations is also deferred to keep this work
incremental and reviewable.