SERVER-86748 Fix maxdiff to sort on difference in area between buckets (#24384)
# Issue

[SERVER-86748](https://jira.mongodb.org/browse/SERVER-86748)

# Description

MaxDiff implementation used sorting on area to determine bucket
boundaries. However, MaxDiff histograms sort on differences in areas
between buckets to find boundaries. Changes were made to the code to
account for this and modify the interface to allow the caller to choose
whether boundaries should be found by sorting on area or areaDiff (with
areaDiff being the default since we are following the implementation of
the MaxDiff algorithm). Unit tests were reused from
maxdiff_histogram_test.cpp and called using both area and areaDiff as
sort values to check if the estimates would be the same or better in
case of areaDiff. Even though, using areaDiff to determine bucket
boundaries provides worse cardinality estimates for some of the specific
cases in the given unit tests, additional testing on different mixed
data distributions and data sizes shows that the error for areaDiff is
significantly low as well.

Minor bug fix was made for the random data generator for normal
distribution.

# Testing

-
https://spruce.mongodb.com/version/669edf1b03474c00077a8843/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC

---------

Co-authored-by: Sopho Kevlishvili <sopho.kevlishvili@mongodb.com>