SERVER-92736 Update symbol mapper to use the original executed task (#25178)
In https://jira.mongodb.org/browse/BF-31405 the symbol mapper will fail
if any of the dependent tasks (archive, push, compile) are restarted and
then when the rested execution is not completed by the time the symbol
mapper is run.

The symbol mapper checks for the success of any of the execution tasks,
failing if any of them are not listed as "success". Since an incomplete
task would not have a status of success, the check fails.

To address this, change the symbol mapper's logic to depend on the first
execution of each of the dependent tasks.

This prevents the symbol mapper from getting a mix of new and old tasks,
while retaining the least strange behavior of defaulting to the original
execution.