SERVER-92668 Fix wrong result with SBE, hash index, $lookup (#25158)
sbe_stage_builder_lookup.cpp buildIndexJoinLookupStage() was incorrectly
assuming that if the index being used has type INDEX_HASHED, then the
foreignField being looked up in that index is the hashed field, but that
is not always the case because a hashed index has exactly ONE hashed
field, so if the index has more than one field in the key, a different
one might be hashed. The failure that resulted was adding an extra step
to compute the hash of the foreignKey value before looking it up in the
index. In the case where that field is not the hashed one, this gave
incorrect query results.

The fix is to hash the value only if foreignField is indeed the hashed
field in the index.