SERVER-89435: Emit a log when validation exceeds maxValidateMemoryUsageMB (#25195)
Currently, such a message is added to the validate command results
object. However, the triage team would like to be able to rely more on
the logs rather than the results object — this would add a log line that
indicates that we have exceeded maxValidateMemoryUsageMB and will not be
able to list all index inconsistencies.