SERVER-92738 use pip to install cryptography 2.3 (#25246)
Poetry tries to be ideal and enforce pep517 to make the building of the
wheel isolated, but I don't think this will work with the version of
cryptography. When doing the same install with pip, it successfully
builds the shared libraries. When letting poetry build the wheel in its
isolated environment, the shared libraries are not built.

This change makes a special case for these platforms and after poetry
has finished its install, go behind poetries back, uninstall the
malformed cryptography package, and reinstall it with pip directly.