SERVER-92804 upload workstation unittest binaries and improve nightly unittest coverage (#25285)
In [BF-33810](https://jira.mongodb.org/browse/BF-33810), we found that
some unittests were running and the binaries and debug symbols were not
being saved, when a error occurred, the core dump was unusable because
the original binaries were lost.

This PR uploads the missing binaries and adds coverage to the nightlies.