SERVER-92181: Retry dropping unfinished indexes during RTS (#25092)
# Issue

SERVER-92181

# Description

According to WT engineers: "drop" will transitively EBUSY sometimes, and
it's not feasible to guarantee it doesn't. For this reason we will retry
the drop.

Accordingly, revert a previous attempt to workaround this issue by
adding wait flags to the WT drop request. Not only does WT not
necessarily honour these waits, but they would prefer we didn't use them
as they're an undocumented/unsupported API.