SERVER-91868 Upload coveralls data (#25273)
Lots of iterations to finally let the dust settle on a simple
implementation.

We don't need to combine the report ourselves to send one data file to
coveralls.
AFAIK, we don't need to send any "done" message to coveralls to say "now
you can merge everything we gave you in a batch of uploads".

Coveralls combines these for us out-of-the-box.
https://coveralls.io/github/10gen/mongo

There is more discovery and polish to be had, but this is a significant
milestone that stands on its own.