SERVER-92685: Update _shardsvrCoordinateMultiUpdate system user authenticated run (#25079)
This PR enhances the _shardsvrCoordinateMultiUpdate auth test in the
following way:

For a user without ['__system'] role, _shardsvrCoordinateMultiUpdate
will return an Unauthorized user error.
For a user with ['__system'] role, _shardsvrCoordinateMultiUpdate should
not return an Unauthorized user error.
In an effort to run this test the following changes are added

sys and norole users are now constant (this is needed for different
commands to have admin as user)
commands can have optional precommands to run. This precommands can be
run on a different server or db.
commands can have optional postcommands to clean up blocking work from a
command.