SERVER-65316 Do not exclude tests from 'inMemory' variant when burn-in fails with WT_CACHE_FULL (#24786)
Some tests run out of memory (WT_CACHE_FULL error) in burn-in mode with
the 'inMemory' engine even when they drop databases and collections they
create. This is a known problem and it can be ignored until SERVER-92452
is fixed. Particularly, these tests should not be completely excluded
from running with the 'inMemory' engine. In this change we add a comment
to the tests which are known to have this issue and removed the
'requires_persistence' tag from the tests that seem to work with the
'inMemory' engine outside of burn-in.