SERVER-92988 Upload bazel unit test binaries before running the unit tests (#25503)
The unit test timeout in https://jira.mongodb.org/browse/BF-34341 cannot
be diagnosed because the timeout is hit before the binaries are
uploaded, making the coredump useless.

Instead move the upload step in between the actual build and the unit
test run, to make sure the binaries will be available in case the unit
tests timeout.