SERVER-82276:  change stream test to demonstrate a working source-merge cycle. (#25379)
The ticket is about "cycle" with mongo-to-mongo not working when we
update mongo sink with "merge". This test demonstrates a cycle with
changestream source config of "fullDocument: updateLookup". The
changestream is not returning fullDocument if "fullDocument" config has
any other value. This does not fix the bug itself but adds a test to be
able to debug the bug.