SERVER-92601 Remove the package_new_compile task from the critical path (#25015)
This PR removes the `package_new_compile` task and replaces it with a
`package_supplementary_data` task that is not on the critical path. This
task creates an archive that contains the supplementary data used by DSI
tests (mongo tools, feature flag list, server params list, jstests, and
buildscripts).

This DSI PR adjusts the download logic for the binaries to directly grab
the output of the compile tasks and adds a separate step to download the
supplementary data:
https://github.com/10gen/dsi/pull/1829