SERVER-92828 Introduce db/commands/query_cmd directory owned by Query Team (#25410)
This change helps to clarify which commands files are owned by the query
team. The movement of source code is a mechanical change that changes
the directory structure but does not change the structure of the libdeps
graph. This patch introduces an OWNERS.yml file in the new query_cmd
directory which avoids adding additional complexity to the pre-existing
owners file in the db/commands directory.