SERVER-93117: Fix small idleTimeout late-DLQ events with change stream $source (#25661)
This PR changes the change stream $source to only send kIdle messages
after receiving an "no events" response from the change stream server.

* One of our customers uses a 200ms window with a 200ms idleTimeout with
a change stream $source. The $source does not have timeField set, and
documents are assigned timestamps using wallTime in the change event
data.
* Before this change, such a small idleTimeout could lead to DLQ events
due to "late arriving data". This is more likely if/when we decrease the
executor's `sourceIdleSleepDurationMs`.
* Prior to this change, a change stream $source could send idle
messages, even if it was in the middle of making a getMore request. This
could lead to a window getting closed, and then shortly afterwards, and
event arriving for that window, which is DLQ-ed.
* The fix is to only send idle messages from the change stream $source
once the background thread has received an empty response from the
changestream server.
* This fix does not help with Kafka $source. librdkafka does not make it
easy to distinguish between an actual idle broker, a long network
request to the broker, and a broken connection with the broker. We will
improve this in https://jira.mongodb.org/browse/SERVER-89760.