SERVER-41314 Make txn_two_phase_commit_failover.js robust to a participant voting abort before other prepare requests are sent (#25576)
Prior to this commit, txn_two_phase_commit_failover.js used the "skip"
field for the failpoints hangWhileTargetingLocalHost and
hangWhileTargetingRemoteHost so that these failpoints wouldn't interfere
with prepare requests from the TransactionCoordinator when the test was
only trying to intercept decision (i.e. abort/commit) requests. It
expected that we would see two prepare requests to the remote to skip,
and 1 to the local process.

However, this method was flaky: if the coordinator learns that a
participant aborted before all of the prepares were sent, it will skip
sending the unsent prepares. In this case, no prepare requests trigger
the failpoints, so our "skip" value is incorrect and we end up hanging
waiting on the failpoint.

To fix this, this commit augments the failpoints with data specifying
which phase of two-phase-commit the test is trying to pause, and only
enters the failpoints if the request is from the appropriate phase. This
means the skips are no longer required.