SERVER-92802 Disable UserCacheInvalidator on ldap_runtime_configuration.js (#25606)
`runChangeLDAPQueryBindTest() `will fail if the `UserCacheInvalidation`
job runs while the lookup call to the config server from mongos is being
made. This is because the call will be invalidated, and a reattempt will
use the new invalid LDAP parameters.

We set `userCacheInvalidationIntervalSecs` to the maximum value to
prevent this job from occurring during the test. Also adds a log to make
this debugging easier in the future.