SERVER-92332 Prevent restoring from yield a non existent collection (#24812)
In 8.0, SBE can produce non-EOF plans for aggregations running on a
non-existing collection (SERVER-92359). Turns out that, under this
circumstances, AutoGet won't stash a CollectionCatalog when recovering
from yield in a lock-free read. This PR adds a tassert to AutoGet for
the case where a non existent collection is restored in the presence of
requested secondary namespaces in a lock-free read.

---------

Co-authored-by: Mihai Andrei <mihai.andrei@mongodb.com>