SERVER-92780: Use DSI based sys-perf configuration (#25703)
Switches to using the sys-perf configuration in the DSI repository.

All the sys-perf variants and tasks have been moved to the DSI
repository, leaving only the compiles and compile-related variants/task
in the mongo repo, along with a reference to the variants and tasks that
should be run in the DSI repository.