SERVER-93213 Optimize getTimeseriesBucketsMayHaveMixedSchemaData to parse the storage engine option only once (#25844)
**Context:**

Our performance test showed that SERVER-91195 is causing a regression of
15%.

After an investigation, we have concluded that the changes for
getTimeseriesBucketsMayHaveMixedSchemaData are responsible for the
regression. The method runs for every time-series CRUD operations, and
it had been changed to get the value from a flag set in the
storageEngine bson object at every iteration, instead of returning the
cached in memory value that could be eventually wrong.

Our local benchmark showed the old implementation was around ~200x times
faster.

**Solution:**

The parsing of the document could be done only once in the constructor
of CollectionImpl and then cached as before. This should both fix
SERVER-91195 and revert the regression.