SERVER-91911 Validate the inputs to $rankFusion (#25905)
Added validation that all input pipelines to $rankFusion are ranked
pipelines. In order to check that these pipeline are also 'Selection
pipelines' and do not modify the documents in any way, added a
constraint 'noFieldModifications' that's set to false by default and
toggled on for stages that are selection-safe (eg. limit and sort).