SERVER-60823: Restructure transaction override retries (#24897)
Previously, the transaction override would retry in a recursive manner;
on failure it would call "runCommand" for each op in the transaction,
which would _again_ enter the transaction override.

It also enforced no timeouts or retry limits.

As a result, an ongoing issue on the server (potentially intentionally
introduced by the test) would cause the txn override to recurse
infinitely - each failed op would deepen the stack.

Eventually, this will hit stack limits, kiling the test.

This has been restructured into an iterative approach, with interation
and time limits.