SERVER-93258 Autopin RBE Dockerfiles (#25885)
There are still various system dependencies missing in the RBE
containers. This commit makes it easier to adjust them all at once while
improving reproducibility and container size.

The Dockerfiles are now generated by a script which may be invoked
manually or via Bazel. The script autodetects the latest sha256 hashes
of the images and determines the fixed package versions. This minimizes
unintentional dependency changes and makes it easier for security
tooling to find vulnerabilities. All Dockerfiles now pass hadolint.

Adjustments to the image contents are now unified in the
`repin_dockerfiles.sh` script. This makes it easier to add dependencies
globally and keep the different images in sync.

Fixes SERVER-93258, SERVER-93257 and all other missing libraries.