SERVER-92337 Reset variables for $unionWith during second execution of subpipeine for explain (#26029)
This PR fixes the issue of the second execution of the subpipeline for
`$unionWith` erroring due to `SEARCH_META` being set.

To do this, the `variables` and `variableParseState` are stored at the
beginning of the first subpipeline execution and restored before the
second subpipeline execution. This follows how other stages, such as
`$lookup` restore variables for multiple executions.

(cherry picked from commit b81561581dfd9489053111902af9a3efdcb579e7)