SERVER-93368 Delete uses of sbe_abt_test_utils library and uses (#25902)
This in preparation for deleting all uses of OptPhaseManager, one of the
main entry points to Bonsai. In order to do so, we need remove all uses
of it in tests. I chose sbe_abt_test_utils as a starting point as it
because it drives many unit tests which take an MQL pipeline, translate
to ABT, optimize via Bonsai, lower to SBE and execute the plan.