SERVER-93400 guard against collmod timing in timeseries_insert_mixed_schema_bucket (#25962)
A BF occurred when a document with mixed schema data was written to a
collection with `timeseriesBucketsMayHaveMixedSchemaData=true` , then
deleted, then `collMod` was called setting
`timeseriesBucketsMayHaveMixedSchemaData=false`

While logically sound, functionally this can lead to a period where that
document still exists at the checkpoint timestamp (which is relevant for
background validation), but
`timeseriesBucketsMayHaveMixedSchemaData=false` since
`timeseriesBucketsMayHaveMixedSchemaData` is stored in an untimestamped
WT table.

This is unlikely to be a customer-facing issue since background
validation is not generally used and this sequence of events is not
really done in practice. This change fixes this from a test standpoint
by calling `fsync` and forcing a checkpoint before the background
validation could find a conflicting time window. A ticket has been filed
so we can discuss in triage if we want to address this validation
behavior.