SERVER-92255: Make KafkaConsumerOperator track maxMemoryUsageBytes correctly. (#26178)
Changes in this pr:
- Reset memory usage before recomputing it in
`KafkaConsumerOperator::doGetStats()`.
- Undo changes I made related to handling of `_memoryUsageHandle` in
KafkaPartitionConsumer in https://github.com/10gen/mongo/pull/25997. I
did not realize that `_memoryUsageHandle` is not thread-safe and that it
needs to be accessed while `_finalizedDocBatch.mutex` is acquired. 🤦
- Acquire mutex in KafkaPartitionConsumer::doGetStats before accessing
`_memoryUsageHandle`.
- Add a `isDocVecFull()` method to de-duplicate a tiny peace of logic
written in 2 places to avoid bugs related to it. This one is a no-op
code cleanup.