SERVER-93497 Defer user cache invalidation to onCommit handler (#26162)
We defer invalidation until after the commit occurs. The way it
currently works, there is a brief window between invalidation and commit
when if a parallel thread reads from the cache, they will read-through
and populate the cache with old data (since the new data has not been
committed). This leads to stale data sitting in the cache until the next
invalidation, which is not guaranteed to ever occur.

I have filed the related ticket,
https://jira.mongodb.org/browse/SERVER-93616, to add more testing
surrounding this. For now, I'd like to prioritize getting this piece
through and fixing the bug before we think about test improvements.