SERVER-93216 Add query stats test for queries that cross the 16MB bson size limit (#26030)
Add a test to verify that enabling query stats does not impact success
of queries that return documents at or just above the 16MB size limit
threshold.

This is particularly relevant in sharded environments where we add a
couple hundred bytes of additional metrics to the response that we send
from mongod to mongos. The addition of these metrics must not cause
valid queries (i.e. queries that respect the 16MB size limit) to fail.

This test is in the core suite, so it will run in the query stats
sharded and standalone passthroughs.