SERVER-93571: Don't throw InvalidOptions in change stream $source watch error handling (#26187)
In this
[incident](https://docs.google.com/document/d/1Dz8RKMwbakSDOaYqq_iATgH0vyQLnW315gP3Tgl2v2k/edit?pli=1#heading=h.my8l8kwcujpo),
we misclassified a change stream $source connectivity error as a non
retryable error, which put a customer processor in a failed state. This
PR fixes the issue. It removes the special handling to classify watch()
exceptions as `StreamProcessorInvalidOptions` (a non-retryable error),
when the user sets $source.config.pipeline. The current approach is too
risky, I will add this back in SERVER-93627 with a different approach.