SERVER-82228: Refactor run_aggregate with AggregationExecutionState class (#25450)
[SERVER-82228](https://jira.mongodb.org/browse/SERVER-82228)

***Description***
This ticket is the first part of a series to refactor the
`run_aggregate.cpp` file, which holds the logic to execute an aggregate
command. The logic in this file is very convoluted and difficult to
follow or edit. This first step is the foundation of the refactor, which
is to pull out common state variables that are needed throughout the
execution. This provides multiple advantages:

1. Its clear to the reader which core state is central to the processing
of an aggregation, independent of the logic. Before, a non-expert would
have trouble deciphering which variables were central, and which ones
were temporary.
2. The sub-functions to `_runAggregate` now have a much shorter
parameter list, and it becomes clearer which variables are distinct to
each function, other than AggregationExecutionState.
3. Going forward, `_runAggregate` will be much easier to further break
into more modular sub-functions, because the majority of the state
necessary for each logical step is encapsulated in
AggregationExecutionState, and thus can be easily factored out, without
a bunch of ties to the state in the top-level funciton.

One note about the philosophy of which state made it into
AggreagtionExecutionState and what didn't. As is always the case in
software design, we want to keep any variable alive during the lifetime
it is needed, and not have it around when its no longer needed. We want
to minimize the amount of variables hanging around in the class that no
longer get used, or invalid placeholders that get set later
(unfortunately this isn't completely avoidable with the view setting
case). So, the idea is that all the variables that should exist for the
lifetime of the aggregation should be in the class, and ones that come
later, or go early should not; and rather be on the stack in
`_runAggregate`. Similarly the member functions of
AggregationExecutionState should rely entirely, or almost entirely, on
the state held within the class and should answer a clear question or
perform an encapsulated task. There is some arbitrary nature to what
becomes a member function, and what is a sub-function to `_runAggregate`
with AggregationExecutionState passed in a a parameter by reference.

Note: lots of credit here is due to @wbuerger46. He came up with the
initial roadmap for the refactor, and took a
[stab](https://github.com/10gen/mongo/tree/will.buerger/SERVER-82228) at
executing most of it during a sunkworks, which this PR took lots of
guidance from - especially in the beginning.

***Testing***
Full evergreen runs are used here to validate that no regressions have
been introduced. Luckily this is a very central code path, so a large
number of the tests in the suite run through this logic, and thus our
coverage here is very high. I can attest to this as even the smallest
errors in the code would show up as failures in evergreen while
developing.

Performance testing was also completed to show this change does not introduce any regressions.

---------

Co-authored-by: Ted Tuckman <TedTuckman@users.noreply.github.com>
Co-authored-by: Charlie Swanson <charlie.swanson@mongodb.com>
Co-authored-by: Will Buerger <59492746+wbuerger46@users.noreply.github.com>