SERVER-90398 Fixed the issue with the error message when running count using _id as key (#22961)
# Issue

[SERVER-90398](https://jira.mongodb.org/browse/SERVER-90398)

# Description

Using "_id" as key in $count produces a misleading error message saying
that _id can only be specified once. To fix this problem,
document_source_count.cpp was modified such that when "_id" is used as a
key, the error message produced says "the count field cannot be '_id'.
This was done by adding another uassert function to the file which
checks if the key is equal to "_id".

Change was made to the document_source_count_test.cpp file as well. An
additional unit test was added to check for the issue above.

---------

Co-authored-by: Sopho Kevlishvili <sopho.kevlishvili@mongodb.com>
Co-authored-by: Chi-I Huang <chii.huang@mongodb.com>