SERVER-90741 optimizing nor+alwaysTrue/alwaysFalse (#23345)
# Issue

[SERVER-90741](https://jira.mongodb.org/browse/SERVER-90741)

# Description

We saw that $nor + $alwaysTrue and $nor + $alwaysFalse were not
optimized in all contexts to $alwaysFalse and $alwaysTrue, respectively.
The issue was caused by invalid conversion from $nor to $not in cases
such as above. To account for this, expression_tree.cpp was modified to
add additional conditions for conversion from $nor to $not so that the
conversion occurs only for valid predicates. This ensures that
$nor+$alwaysTrue/alwaysFalse is not converted to
$not+$alwaysTrue/alwaysFalse and therefore, does not result in an error
causing these expressions being left un-optimized.

Unit tests were added to expression_simplifier_test.cpp to test for the
behavior above. Jstests were added to assert that EOF plan is generated
when an expression gets simplified to $alwaysFalse (e.g. for
$nor+$alwaysTrue).

---------

Co-authored-by: Sopho Kevlishvili <sopho.kevlishvili@mongodb.com>
Co-authored-by: Chi-I Huang <chii.huang@mongodb.com>