SERVER-93620 Fix SBE stage trackResult (#26203)
In SERVER-88047 (#24922), we remove `TrialRunTrackingType::TrackResults`
in `doAttachToTrialRunTracker()` function. But after that we are not
able to detach the tracker when reach max kResult metric and upgrade to
a normal run, this causes trackRead() function to throw an exception
with early exit.

This causes unnecessary replanning and performance regression for some
SBE queries.