SERVER-93567: Return stats in notifyFlush response (#26295)
This PR will allow the streams Agent to save the stats of a processor
during pod shutdown (i.e. deployments).

In https://jira.mongodb.org/browse/STREAMS-1081 we implemented support
for returning stats when the processor is in a stopped state. The stats
are stored on the metadataDB.
* Currently the Agent only updates the stats in metadataDB during a
user-initiated stop request.
* We also want to update the stats in the metadataDB during a planned
pod shutdown. Since mongostream shuts itself down on its own, the
easiest place for Agent to get the stats during shutdown is the response
of its `notify checkpoint flushed` call. mongostream already waits on
this call before stopping each stream [processor.