SERVER-93608 Move Locker out of wiredtigerPrepareConflictRetry (#26157)
Implements APIs on the `RecoveryUnit` and `Locker` to support the
assertions causing the layering violations.

Uses the new APIs in the oplog allocation path and in the DB/Coll RAII
helpers.

I opted to keep the assertions the same in the Locker rather than trying
to build an API that was less process-fatal. That could be a good
follow-up for CAR.