SERVER-93619 Remove ABT lowering dependency on optimizer Metadata and NodeToGroupPropsMap (#26274)
This patch is one of several to unravel the dependency edge between ABT
lowering code using by SBE stage builders with optimizer concepts such
as Metadata. We're moving towards deleting the query/optimizer directory
entirely and the ABT lowering code is one of the few remaining libraries
that depends on the optimizer.

---------

Co-authored-by: Mihai Andrei <mihai.andrei@mongodb.com>