SERVER-93435 Do not convert StaleConfig for updateMany in a transaction to QueryPlanKilled (#26422)
SERVER-91465 made a change that converted `StaleConfig` errors to
`QueryPlanKilled` in updateMany commands in the presence of migrations.
Returning `StaleConfig` made the router retry, leading to potentially
incorrect results. This however is not necessary in case the write is
part of a transaction, since it is safe to retry a transaction anyway,
and it led to other problems (BF-34413). This PR prevents such
conversion when in a multidocument transaction.