SERVER-93921 allow DEVELOPER_DIR to be set from environment (#26474)
DEVELOPER_DIR was hardcoded in the config.h header. this however can
causes issues if the hardcoded developer dir happens to not exist on the
target system, such as the case for developers macbooks. This change
makes it so the DEVELOPER_DIR can be set from the environment via scons
(and then on bazel command line).